backend/requirements.txt
openai==0.28
flask
pymongo
python-dotenv
